Move grdimage function to grdimage.py
