+++ pygmt/src/__init__.py
@@ -7,6 +7,7 @@ from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
 from pygmt.src.contour import contour
 from pygmt.src.grdcontour import grdcontour
+from pygmt.src.grdimage import grdimage
 from pygmt.src.grdview import grdview
 from pygmt.src.image import image
 from pygmt.src.legend import legend
